// client/src/js/bindings/LoginResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Keypair } from "./Keypair";
import type { ServerPermissions } from "./ServerPermissions";

export type LoginResponse = { gamerpic: string, gamertag: string, keypair: Keypair, signature: Keypair, certificate: string, certificate_key: string, certificate_ca: string, quic_connect_string: string, minecraft_username: string | null, server_permissions: ServerPermissions | null, };